requirements.txt
pathlib
python-dotenv
matplotlib
pydicom